Uldate

+++ Server/Models/StatFunctionEnvironment.cs
@@ -4,35 +4,35 @@
     {
         public CharacterEntity Character { get; set; }
 
-        public float Value { get; set; }
-        public float BaseValue { get; set; }
+        public double Value { get; set; }
+        public double BaseValue { get; set; }
 
-        public void AddValue(float value)
+        public void AddValue(double value)
         {
             Value += value;
         }
 
-        public void SubValue(float value)
+        public void SubValue(double value)
         {
             Value -= value;
         }
 
-        public void MulValue(float value)
+        public void MulValue(double value)
         {
             Value *= value;
         }
 
-        public void DivValue(float value)
+        public void DivValue(double value)
         {
             Value /= value;
         }
 
-        public void AddPrecent(float value)
+        public void AddPrecent(double value)
         {
             Value += Value / 100 * value;
         }
 
-        public void SubPrecent(float value)
+        public void SubPrecent(double value)
         {
             Value -= Value / 100 * value;
         }
